# public/swagger-specs/services/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ConsenSys Codefi Orchestrate API. Enables dynamic management of transactions, identities, chains, faucets and contracts.\nTransaction Requests are an abstraction over schedules and jobs representing one or more transactions executed on the Blockchain network\nSchedules are ordered lists of jobs executed in a predefined sequence\nJobs represent single transaction flows executed on the Blockchain network\nChains represent list of endpoints pointing to a Blockchain network\nFaucets represent funded accounts (holding ETH) linked to specific chains, allowed to fund newly created accounts automatically for them to be able to send transactions.\nAccounts represent Ethereum accounts (private keys). By usage of the generated cryptographic key pair, accounts can be used to sign/verify and to encrypt/decrypt messages.\nContracts represent Solidity contracts management.\nEvent Streams represent Event streams management.",
        "title": "Orchestrate API",
        "contact": {
            "name": "Contact ConsenSys Codefi Orchestrate",
            "url": "https://consensys.net/codefi/orchestrate/contact",
            "email": "orchestrate@consensys.net"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "paths": {
...
        "/chains": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chains"
                ],
                "summary": "Retrieves a list of all registered chains",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "allOf": [
                                    {
                                        "$ref": "#/definitions/types.ChainResponse"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "privateTxManager": {
                                                "$ref": "#/definitions/entities.PrivateTxManager"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chains"
                ],
                "summary": "Registers a new chain",
                "parameters": [
                    {
                        "description": "Chain registration request.",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.RegisterChainRequest"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "listener": {
                                            "$ref": "#/definitions/types.RegisterListenerRequest"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.ChainResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "privateTxManager": {
                                            "$ref": "#/definitions/entities.PrivateTxManager"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/chains/{uuid}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chains"
                ],
                "summary": "Retrieves a chain by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the chain",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ChainResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Chain not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chains"
                ],
                "summary": "Deletes a chain by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the chain",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Chain not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chains"
                ],
                "summary": "Updates a chain by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the chain",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Chain update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.UpdateChainRequest"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "listener": {
                                            "$ref": "#/definitions/types.UpdateListenerRequest"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ChainResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Chain not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 24000
                },
                "message": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "entities.PrivateTxManager": {
            "type": "object",
            "properties": {
                "chainUUID": {
                    "description": "UUID of the registered chain.",
                    "type": "string"
                },
                "createdAt": {
                    "description": "Date and time that the private transaction manager was registered with the chain.",
                    "type": "string"
                },
                "type": {
                    "description": "Currently supports `Tessera` and `EEA`.",
                    "type": "string"
                },
                "url": {
                    "description": "Transaction manager endpoint.",
                    "type": "string"
                },
                "uuid": {
                    "description": "UUID of the private transaction manager.",
                    "type": "string"
                }
            }
        },
...
        "types.ChainResponse": {
            "type": "object",
            "properties": {
                "chainID": {
                    "description": "[Ethereum chain ID](https://besu.hyperledger.org/en/latest/Concepts/NetworkID-And-ChainID/).",
                    "type": "integer",
                    "example": 2445
                },
                "createdAt": {
                    "description": "Date and time at which the chain was registered.",
                    "type": "string",
                    "example": "2020-07-09T12:35:42.115395Z"
                },
                "labels": {
                    "description": "List of custom labels.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "listenerBackOffDuration": {
                    "description": "Time to wait before trying to fetch a new mined block.",
                    "type": "string",
                    "example": "5s"
                },
                "listenerCurrentBlock": {
                    "description": "Current block.",
                    "type": "integer",
                    "example": 0
                },
                "listenerDepth": {
                    "description": "Block depth after which the Transaction Listener considers a block final and processes it.",
                    "type": "integer",
                    "example": 0
                },
                "listenerExternalTxEnabled": {
                    "description": "Whether the chain listens for external transactions not crafted by Orchestrate.",
                    "type": "boolean",
                    "example": false
                },
                "listenerStartingBlock": {
                    "description": "Block at which the Transaction Listener starts processing transactions",
                    "type": "integer",
                    "example": 5000
                },
                "name": {
                    "description": "Name of the chain.",
                    "type": "string",
                    "example": "mainnet"
                },
                "ownerID": {
                    "description": "ID of the chain owner.",
                    "type": "string",
                    "example": "foo"
                },
                "privateTxManagerURL": {
                    "description": "Private tx manager required by go-quorum for sending private txs",
                    "type": "string",
                    "example": "http://go-quorum/tessera:9000"
                },
                "tenantID": {
                    "description": "ID of the tenant executing the API.",
                    "type": "string",
                    "example": "tenant"
                },
                "updatedAt": {
                    "description": "Date and time at which the chain details were updated.",
                    "type": "string",
                    "example": "2020-07-09T12:35:42.115395Z"
                },
                "urls": {
                    "description": "URLs of Ethereum nodes connected to.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "https://mainnet.infura.io/v3/a73136601e6f4924a0baa4ed880b535e"
                    ]
                },
                "uuid": {
                    "description": "UUID of the registered chain.",
                    "type": "string",
                    "example": "b4374e6f-b28a-4bad-b4fe-bda36eaf849c"
                }
            }
        },
...
        "types.RegisterChainRequest": {
            "type": "object",
            "required": [
                "name",
                "urls"
            ],
            "properties": {
                "labels": {
                    "description": "List of custom labels. Useful for adding custom information to the chain.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "listener": {
                    "$ref": "#/definitions/types.RegisterListenerRequest"
                },
                "name": {
                    "description": "Name of the chain. Must be unique.",
                    "type": "string",
                    "example": "mainnet"
                },
                "privateTxManagerURL": {
                    "description": "Private tx manager required by go-quorum for sending private txs",
                    "type": "string",
                    "example": "http://go-quorum/tessera:9000"
                },
                "urls": {
                    "description": "List of URLs of Ethereum nodes to connect to.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "https://mainnet.infura.io/v3/a73136601e6f4924a0baa4ed880b535e"
                    ]
                }
            }
        },
...
        "types.RegisterListenerRequest": {
            "type": "object",
            "properties": {
                "backOffDuration": {
                    "description": "Time to wait before trying to fetch a new mined block (for example `1s` or `1m`, default is `5s`).",
                    "type": "string",
                    "example": "1s"
                },
                "depth": {
                    "description": "Block depth after which the Transaction Listener considers a block final and processes it (default 0).",
                    "type": "integer",
                    "example": 0
                },
                "externalTxEnabled": {
                    "description": "Whether to listen to external transactions not crafted by Orchestrate (default `false`).",
                    "type": "boolean",
                    "example": false
                },
                "fromBlock": {
                    "description": "Block from which the Transaction Listener should start processing transactions (default `latest`).",
                    "type": "string",
                    "example": "latest"
                }
            }
        },
...
        "types.UpdateChainRequest": {
            "type": "object",
            "properties": {
                "labels": {
                    "description": "List of custom labels. Useful for adding custom information to the chain.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "listener": {
                    "$ref": "#/definitions/types.UpdateListenerRequest"
                },
                "name": {
                    "description": "Name of the chain. Must be unique.",
                    "type": "string",
                    "example": "mainnet"
                }
            }
        },
...
        "types.UpdateListenerRequest": {
            "type": "object",
            "properties": {
                "backOffDuration": {
                    "description": "Time to wait before trying to fetch a new mined block (for example `1s` or `1m`, default is `5s`).",
                    "type": "string",
                    "example": "1s"
                },
                "currentBlock": {
                    "description": "Latest block number fetched.",
                    "type": "integer",
                    "example": 1
                },
                "depth": {
                    "description": "Block depth after which the Transaction Listener considers a block final and processes it (default 0).",
                    "type": "integer",
                    "example": 0
                },
                "externalTxEnabled": {
                    "description": "Whether to listen to external transactions not crafted by Orchestrate (default `false`).",
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "JWTAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}